Moved resolve_reference() to utils
